+++ Assets/Scripts/Game/PauseController.cs
@@ -6,10 +6,20 @@ public class PauseController : MonoBehaviour
 {
     public GameObject pauseCanvas;
     private bool isPaused = false;
+    private bool isPausing = false;
+    private bool isDepausing = false;
     private bool canPause = false;
+    public Fader overlay;
+    private AudioManager audioManager;
+
+    void Nil()
+    {
+    }
 
 	void Start ()
     {
+        Debug.Log (Time.fixedDeltaTime);
+        audioManager = GameObject.Find("AudioManager").GetComponent<AudioManager>();
         pauseCanvas.SetActive (false);
         Invoke ("PauseActive", 2.0f);
 	}
@@ -32,6 +42,7 @@ public class PauseController : MonoBehaviour
 
         if (isPaused && Input.GetButtonDown("Cancel")) {
             Time.timeScale = 1;
+            Time.fixedDeltaTime = 0.02F;
             SceneManager.LoadScene ("Title");
         }
 
@@ -40,16 +51,57 @@ public class PauseController : MonoBehaviour
         }
 	}
 
+    void FixedUpdate()
+    {
+        if (isPausing) {
+            audioManager.musicSource.pitch = Mathf.Lerp(audioManager.musicSource.pitch, 0.5f, 0.05f);
+            Time.timeScale = Mathf.Lerp (Time.timeScale, 0, 0.07f);
+            Time.fixedDeltaTime = 0.02F * Time.timeScale;
+            if (Time.timeScale < 0.001f) {
+                Time.timeScale = 0;
+                isPausing = false;
+                Time.fixedDeltaTime = 0.02F;
+            }
+        }
+
+        if (isDepausing) {
+            audioManager.musicSource.pitch = Mathf.Lerp(audioManager.musicSource.pitch, 1f, 0.2f);
+            Time.timeScale = Mathf.Lerp (Time.timeScale, 1, 0.2f);
+            Time.fixedDeltaTime = 0.02F * Time.timeScale;
+
+            if (Time.timeScale > 0.99f) {
+                Time.timeScale = 1;
+                isDepausing = false;
+                Time.fixedDeltaTime = 0.02F;
+            }
+        }
+    }
+
     void TogglePause()
     {
         isPaused = !isPaused;
+
         if (isPaused) {
-            pauseCanvas.SetActive(true);
-            Time.timeScale = 0;
+            ActivatePause ();
+            overlay.Fade(new Color(0,0,0,0.5f), 7, Nil);
+            isDepausing = false;
+            isPausing = true;
         } else {
-            pauseCanvas.SetActive(false);
-            Time.timeScale = 1;
+            Time.timeScale = 0.1f;
+            overlay.Fade(new Color(0,0,0,0f), 7, DeactivatePause);
+            isPausing = false;
+            isDepausing = true;
         }
 
+    }
+
+    void ActivatePause()
+    {
+        pauseCanvas.SetActive(true);
+    }
+
+    void DeactivatePause()
+    {
+        pauseCanvas.SetActive(false);
     }
 }